requirements.txt
absl-py==0.7.1
astor==0.8.0
boto==2.49.0
boto3==1.9.214
botocore==1.12.214
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cycler==0.10.0
decorator==4.4.0
docutils==0.15.2
efficientnet==1.0.0
gast==0.2.2
gensim==3.8.0
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.9.0
idna==2.8
imageio==2.5.0
imgaug==0.3.0
jmespath==0.9.4
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lxml==4.4.1
Markdown==3.1.1
matplotlib==3.1.1
networkx==2.3
nltk==3.4.5
numpy==1.17.0
opencv-python==4.1.0.25
opencv-python-headless==4.1.2.30
opt-einsum==3.1.0
pandas==0.25.1
Pillow==6.1.0
protobuf==3.9.1
psutil==5.6.7
pyparsing==2.4.2
python-dateutil==2.8.0
python-docx==0.8.10
pytz==2019.2
PyWavelets==1.0.3
requests==2.22.0
s3transfer==0.2.1
scikit-image==0.15.0
scikit-learn==0.22.1
scipy==1.3.1
Shapely==1.6.4.post2
six==1.12.0
smart-open==1.8.4
tensorflow-gpu==2.1.0
termcolor==1.1.0
Theano==1.0.4
tqdm==4.41.1
urllib3==1.25.3
Werkzeug==0.15.5
wrapt==1.11.2
xdg==4.0.1
transformers==2.5.0
